platform: generic: spacemit: add missing objects.mk
Add the missing objects.mk for the SpacemiT platform, required for the
K1 platform to be included in the build.

Fixes: 1f84ec2a ("platform: generic: spacemit: add K1")
Acked-by: Troy Mitchell <troy.mitchell@linux.spacemit.com>
Link: https://lore.kernel.org/r/20251124-k1-fix-v1-1-8d7e7a29379e@pigmoral.tech

+++ platform/generic/spacemit/k1.c
@@ -51,10 +51,10 @@ static void spacemit_k1_pre_init(void)
 	scratch = sbi_scratch_thishart_ptr();
 
 	writel(scratch->warmboot_addr, (unsigned int *)C0_RVBADDR_LO_ADDR);
-	writel(scratch->warmboot_addr >> 32, (unsigned int *)C0_RVBADDR_HI_ADDR);
+	writel((u64)scratch->warmboot_addr >> 32, (unsigned int *)C0_RVBADDR_HI_ADDR);
 
 	writel(scratch->warmboot_addr, (unsigned int *)C1_RVBADDR_LO_ADDR);
-	writel(scratch->warmboot_addr >> 32, (unsigned int *)C1_RVBADDR_HI_ADDR);
+	writel((u64)scratch->warmboot_addr >> 32, (unsigned int *)C1_RVBADDR_HI_ADDR);
 
 	for (i = 0; i < PLATFORM_MAX_CPUS; i++) {
 		clusterid = CPU_TO_CLUSTER(i);
